Use bipartite graph for the infrastructure and tests (#8)
This commit changes the tests to use https://github.com/ReactiveBayes/BipartiteFactorGraphs.jl. In addition it introduces many building blocks to run inference in the model. It introduces a concept of dependencies between Slots. Every marginal or a message has its own list of dependencies, which can be customized for different algorithms, e.g. for BP, EP or VMP. More inference routines will be added in the next PRs.